Create AdminUserRegisterService by extracting register() from AdminUserAuthService
To improve code modularity and separation of concerns, this commit creates a new service named AdminUserRegisterService and moves the register() method from AdminUserAuthService to the new service. This change allows for more focused responsibilities and better organization of the codebase.

The affected files are:
- AdminUserAuthService.java: removed register() method
- AdminUserRegisterService.java: new service created with register() method
